(qr): Move qr to its own plugin
